# pmcluster/docs/openapi.yaml
openapi: 3.0.3
info:
  title: pmcluster daemon API
  description: |
    REST + webhook API for the pmcluster Docker Swarm control plane.

    All `/api/*` routes require a Bearer token (an API user; see `POST /api/api_keys`
    to mint one). The webhook receiver `/webhook/{source}` is intentionally NOT
    Bearer-authenticated — its auth is an HMAC-SHA256 signature computed over the
    (timestamp + body) with the source's shared secret. API users are minted with
    `pmcluster user create` or `POST /api/api_keys`; webhook sources with
    `pmcluster webhook add` or `POST /api/webhooks`.
  version: "0.2.0"
servers:
  - url: https://pmcluster.example.com
    description: Deployed daemon behind the edge gateway
  - url: http://127.0.0.1:PORT
    description: Local `pmcluster serve`
paths:
...
  /api/stacks:
    get:
      summary: List deployed stacks
      operationId: listStacks
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Stack list
          content:
            application/json:
              schema:
                type: object
                properties:
                  stacks:
                    type: array
                    items:
                      $ref: "#/components/schemas/Stack"
    post:
      summary: Deploy a manifest (create or update a stack)
      operationId: deploy
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeployPayload"
      responses:
        "200":
          description: Deploy accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeployResult"
        "400":
          $ref: "#/components/responses/BadRequest"
  /api/stacks/{name}:
    get:
      summary: One stack's metadata and recent revisions
      operationId: getStack
      security:
        - bearerAuth: []
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Stack detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StackDetail"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/stacks/{name}/revisions/{rev}:
    get:
      summary: One revision's full source and rendered YAML
      operationId: getRevision
      security:
        - bearerAuth: []
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
        - name: rev
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Revision detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Revision"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/stacks/{name}/rollback:
    post:
      summary: Roll a stack back to a prior revision
      operationId: rollback
      security:
        - bearerAuth: []
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [revision]
              properties:
                revision:
                  type: integer
                  format: int64
      responses:
        "200":
          description: Rolled back
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RollbackResult"
        "400":
          $ref: "#/components/responses/BadRequest"
...
  /api/tls/site:
    get:
      summary: Get the cluster's own (default) certificate metadata
      description: |
        Returns metadata for the cluster's own-domain certificate (domain,
        secret names, validity window, SANs, hashes). The PEM bytes themselves
        live in versioned Swarm secrets and are never returned. 404 until a
        certificate has been uploaded.
      operationId: getSiteCert
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Site certificate metadata
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SiteCert"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      summary: Upload/renew the cluster's own cert/key (as PEM text) and refresh Traefik
      description: |
        Cert and key arrive as PEM TEXT in the JSON body — never as a file upload.
        The pair is validated against the cluster's persisted domain, written as
        the local source of truth under ~/.pmcluster/config/site/, materialized as
        versioned Swarm secrets (cert_vN/key_vN), wired into the Traefik dynamic
        config, and its metadata recorded for expiry monitoring.
      operationId: putSiteCert
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [cert, key]
              properties:
                cert:
                  type: string
                key:
                  type: string
      responses:
        "200":
          description: Uploaded and Traefik refreshed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SiteCert"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: pmc_<token_id>_<secret>
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
...
    DeployPayload:
      type: object
      properties:
        app_name:
          type: string
        repo_url:
          type: string
        version:
          type: string
        manifest:
          type: string
      required:
        - manifest
    DeployResult:
      type: object
      properties:
        stack:
          type: string
        revision:
          type: integer
          format: int64
    RollbackResult:
      type: object
      properties:
        stack:
          type: string
        new_revision:
          type: integer
          format: int64
        rolled_back_to:
          type: integer
          format: int64
...
    Stack:
      type: object
      properties:
        name:
          type: string
        current_revision:
          type: integer
          format: int64
        repo_url:
          type: string
        created_at:
          type: integer
          format: int64
        updated_at:
          type: integer
          format: int64
    StackDetail:
      type: object
      properties:
        stack:
          $ref: "#/components/schemas/Stack"
        revisions:
          type: array
          items:
            type: object
            properties:
              revision:
                type: integer
                format: int64
              created_at:
                type: integer
                format: int64
        last_backup:
          $ref: "#/components/schemas/BackupSummary"
    BackupSummary:
      type: object
      properties:
        status:
          type: string
        started_at:
          type: integer
          format: int64
        finished_at:
          type: integer
          format: int64
        revision:
          type: integer
          format: int64
        error_message:
          type: string
    Revision:
      type: object
      properties:
        stack:
          type: string
        revision:
          type: integer
          format: int64
        created_at:
          type: integer
          format: int64
        source_yaml:
          type: string
        rendered_yaml:
          type: string
        payload:
          type: string
...
    SiteCert:
      type: object
      properties:
        domain:
          type: string
        cert_secret:
          type: string
        key_secret:
          type: string
        not_before:
          type: string
          format: date-time
        not_after:
          type: string
          format: date-time
        sans:
          type: array
          items:
            type: string
        cert_hash:
          type: string
        key_hash:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...